Refactor to accomodate 'to' param * Move bulk of code into utils.py as two functions ** `create_suggestion` which takes one instance object of type user or group and returns suggestion code needed by js ** `find_users_and_groups_by_search` will return either all users and groups that match search criteria OR exact match in the case of a `to` param -- this avoids cases where something like `?to=s` would return all items starting with s on a GET request
